+++ Sources/RigControl/RigControl.swift
@@ -12,8 +12,8 @@
 /// ```swift
 /// import RigControl
 ///
-/// let rig = RigController(
-///     radio: .icomIC9700,
+/// let rig = try RigController(
+///     radio: .icomIC9700(),
 ///     connection: .serial(path: "/dev/cu.IC9700", baudRate: 115200)
 /// )
 ///
@@ -30,8 +30,8 @@
 /// ```swift
 /// import RigControl
 ///
-/// let rig = RigController(
-///     radio: .icomIC7600,
+/// let rig = try RigController(
+///     radio: .icomIC7600(),
 ///     connection: .serial(path: "/dev/cu.IC-7600", baudRate: 19200)
 /// )
 /// try await rig.connect()